requirements.txt
opencv_python_headless==4.5.3.56
pdf2image==1.16.0
loguru==0.5.3
pandas==1.3.2
starlette==0.14.2
easyocr==1.4.1
numpy==1.19.5
scipy==1.7.1
torch_snippets==0.463
pydantic==1.8.2
lxml==4.6.3
fastapi==0.68.1
Pillow
sip==6.1.1
aiofiles==0.7.0
uvicorn
python-multipart
